requirements.txt
fastapi
uvicorn[standard]
langchain
langchain_openai
langchain_community>=0.0.10
pypdf
unstructured
python-multipart
tiktoken
python-dotenv
reportlab
pydantic
openai
pdfkit
stripe
sendgrid
requests
reportlab
reportlab==4.0.4
stripe==6.5.0
stripe==6.5.0
email-validator==2.1.0
pydantic[email]>=2.7.0
